SC 117051 Content prep / cleanup (part 3)
