+++ src/applications/appeals/995/pages/evidence/privatePrompt.js
@@ -1,4 +1,4 @@
-import { HAS_PRIVATE_EVIDENCE } from '../constants';
+import { HAS_PRIVATE_EVIDENCE } from '../../constants';
 
 export default {
   uiSchema: {
