+++ src/applications/appeals/995/tests/pages/evidence/privateAuthorization.unit.spec.jsx
@@ -5,13 +5,13 @@ import { render } from '@testing-library/react';
 import { DefinitionTester } from 'platform/testing/unit/schemaform-utils';
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
 
-import formConfig from '../../config/form';
+import formConfig from '../../../config/form';
 
-describe('Supplemental Claims (Custom) Private evidence page', () => {
+describe('Supplemental Claims Private evidence authorization page', () => {
   const {
     schema,
     uiSchema,
-  } = formConfig.chapters.evidence.pages.evidencePrivateRecords;
+  } = formConfig.chapters.evidence.pages.privateAuthorization;
 
   // Custom page is rendered, so this only renders a submit button
   it('should render', () => {
