+++ src/applications/appeals/995/tests/pages/evidence/privateDetails.unit.spec.jsx
@@ -5,15 +5,15 @@ import { render } from '@testing-library/react';
 import { DefinitionTester } from 'platform/testing/unit/schemaform-utils';
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
 
-import formConfig from '../../config/form';
+import formConfig from '../../../config/form';
 
-describe('Supplemental Claims Private evidence request page', () => {
+describe('Supplemental Claims (Custom) Private evidence page', () => {
   const {
     schema,
     uiSchema,
-  } = formConfig.chapters.evidence.pages.evidencePrivateRecordsRequest;
+  } = formConfig.chapters.evidence.pages.privateDetails;
 
-  // Custom page is rendered, so this renders a checkbox + submit button
+  // Custom page is rendered, so this only renders a submit button
   it('should render', () => {
     const { container } = render(
       <DefinitionTester
